requirements.txt
langchain==0.0.284
python-dotenv
streamlit
tiktoken
faiss-cpu==1.7.4
protobuf~=3.19.0
langchain_experimental
mysql-connector-python
pymysql
sentence-transformers
chromadb
langchain-huggingface